moved settings to dashboard
