+++ src/components/Settings/Info/Info.vue
@@ -1,7 +1,6 @@
 <template>
   <v-container>
     <v-card >
-
       <v-card-title>
         Cluster Details
       </v-card-title>
