+++ src/components/Settings/KubernetesSettings/KubeConfig.vue
