+++ src/components/Settings/KubernetesSettings/KubeInstallation.vue
