+++ src/components/Settings/KubernetesSettings/OSInstallation.vue
